+++ domain/src/main/java/co/com/sofkau/clinica/administracion/consultorio/commands/AsignarMedico.java
@@ -1,9 +1,9 @@
-package co.com.sofkau.clinica.administracion.commands;
+package co.com.sofkau.clinica.administracion.consultorio.commands;
 
 import co.com.sofka.domain.generic.Command;
-import co.com.sofkau.clinica.administracion.values.Especialidad;
-import co.com.sofkau.clinica.administracion.values.MedicoId;
-import co.com.sofkau.clinica.administracion.values.Nombre;
+import co.com.sofkau.clinica.administracion.consultorio.commands.values.Especialidad;
+import co.com.sofkau.clinica.administracion.consultorio.commands.values.MedicoId;
+import co.com.sofkau.clinica.administracion.consultorio.commands.values.Nombre;
 
 public class AsignarMedico extends Command {
