+++ domain/src/main/java/co/com/sofkau/clinica/administracion/consultorio/commands/Auxiliar.java
@@ -1,8 +1,8 @@
-package co.com.sofkau.clinica.administracion;
+package co.com.sofkau.clinica.administracion.consultorio.commands;
 
 import co.com.sofka.domain.generic.Entity;
-import co.com.sofkau.clinica.administracion.values.AuxiliarId;
-import co.com.sofkau.clinica.administracion.values.Nombre;
+import co.com.sofkau.clinica.administracion.consultorio.commands.values.AuxiliarId;
+import co.com.sofkau.clinica.administracion.consultorio.commands.values.Nombre;
 
 public class Auxiliar extends Entity<AuxiliarId> {
